graph-ql backend - added readme.md file, renamed files
