+++ metadata_validation_conversion/graphql_api/grapheneObjects/analysis/schema.py
@@ -3,9 +3,9 @@ from graphene.relay import Connection, Node
 from graphql_api.tasks import resolve_all_task
 from celery.result import AsyncResult
 from ..helpers import fetch_index_records, fetch_with_join
-from .fieldObjects import AnalysisDateField, AnalysisJoinField, FilesField, AnalysisOrganismField
+from .field_objects import AnalysisDateField, AnalysisJoinField, FilesField, AnalysisOrganismField
 from .arguments.filter import AnalysisFilterArgument
-from ..commonFieldObjects import ProtocolField, TaskResponse
+from ..common_field_objects import ProtocolField, TaskResponse
 
 
 def fetch_single_analysis(kwargs):
