+++ metadata_validation_conversion/graphql_api/grapheneObjects/experiment/arguments/filter.py
@@ -1,5 +1,5 @@
 from graphene import Field, String, InputObjectType, List
-from ...commonInputFieldObject import ProtocolInputField, UnitInputField
+from ...common_input_field_object import ProtocolInputField, UnitInputField
 
 
 class CustomFieldInputField(InputObjectType):
