+++ metadata_validation_conversion/graphql_api/grapheneObjects/file/schema.py
@@ -3,10 +3,10 @@ from graphene.relay import Connection, Node
 from graphql_api.tasks import resolve_all_task
 from celery.result import AsyncResult
 from ..helpers import fetch_index_records, fetch_with_join
-from .fieldObjects import FileExperimentField, FileJoinField, FilePublishedArticlesField, RunField, SpeciesField, \
+from .field_objects import FileExperimentField, FileJoinField, FilePublishedArticlesField, RunField, SpeciesField, \
     StudyField
 from .arguments.filter import FileFilterArgument
-from ..commonFieldObjects import TaskResponse
+from ..common_field_objects import TaskResponse
 
 
 def fetch_single_file(args):
