+++ metadata_validation_conversion/graphql_api/grapheneObjects/specimen/schema.py
@@ -3,11 +3,11 @@ from graphene.relay import Connection,Node
 from graphql_api.tasks import resolve_all_task
 from celery.result import AsyncResult
 from ..helpers import fetch_index_records, fetch_with_join
-from .fieldObjects import CellCultureField,SpecimenOrganismField,CellLineField,CellSpecimenField,\
+from .field_objects import CellCultureField,SpecimenOrganismField,CellLineField,CellSpecimenField,\
     SpecimenOrganizationField,PoolOfSpecimensField,SpecimenPublishedArticlesField,SpecimenCustomFieldField,\
     SpecimenFromOrganismField,SpecimenJoinField
 from .arguments.filter import SpecimenFilterArgument
-from ..commonFieldObjects import OntologyField, TaskResponse
+from ..common_field_objects import OntologyField, TaskResponse
 
 
 def fetch_single_specimen(args):
